requirements.txt
astroid==2.4.1
certifi==2020.4.5.1
cloudinary==1.20.0
dj-database-url==0.5.0
Django==1.11.29
django-filter==2.2.0
django-forms-bootstrap==3.1.0
django-materializecss-form==1.1.17
django-mathfilters==1.0.0
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.1.2
psycopg2==2.8.5
pylint==2.5.2
pytz==2020.1
six==1.14.0
toml==0.10.0
urllib3==1.25.9
whitenoise==5.0.1
wrapt==1.12.1
